CyberShield_v2/requirements.txt
requests==2.31.0
flask==3.0.0
Pillow==10.2.0
reportlab==4.0.9
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2
reportlab==4.0.9
python-dotenv==1.0.0